Move homepage logic to separate component for multi-page support.
